#124 create variables as needed for models
