+++ pkpdapp/pkpdapp/models/pkpd_model.py
@@ -8,6 +8,7 @@ from django.db import models
 from django.urls import reverse
 from pkpdapp.models import (
     MechanisticModel,
+    MyokitModelMixin,
     Protocol,
     Project,
 )
